requirements.txt
torch
torchvision
transformers
bitsandbytes
Pillow
opencv-python
gradio
groundingdino-py